Made some cleanup in the code. Added more SignalR support. The agent will not ask for activiteid, the Controller Hub will give the agenst tasks to perform.

+++ src/PolymeliaDeploy/Data/Repositories/ActivityRepository.cs
@@ -29,7 +29,7 @@ namespace PolymeliaDeploy.Data.Repositories
         }
 
 
-        public async Task<IEnumerable<ActivityTask>> GetActivityTasks(int lastTaskId, string serverRole)
+        public async Task<IEnumerable<ActivityTask>> GetActivityTasks(long lastTaskId, string serverRole)
         {
             return await Task.Run(() =>
             {
