+++ src/PolymeliaDeploy/Data/Repositories/IActivityRepository.cs
@@ -7,6 +7,6 @@
     {
         Task UpdateActivityTaskStatus(long activityTaskId, ActivityStatus status);
 
-        Task<IEnumerable<ActivityTask>> GetActivityTasks(int lastTaskId, string serverRole);
+        Task<IEnumerable<ActivityTask>> GetActivityTasks(long lastTaskId, string serverRole);
     }
 }
